refactor: create demo data file
