+++ now/apps/image_to_image/app.py
@@ -7,13 +7,8 @@ from now_common.preprocess import preprocess_images
 from now_common.utils import common_setup, get_indexer_config
 
 from now.apps.base.app import JinaNOWApp
-from now.constants import (
-    CLIP_USES,
-    EXTERNAL_CLIP_HOST,
-    Apps,
-    DemoDatasetNames,
-    Modalities,
-)
+from now.constants import CLIP_USES, EXTERNAL_CLIP_HOST, Apps, Modalities
+from now.demo_data import DemoDatasetNames
 from now.now_dataclasses import UserInput
