+++ tests/integration/data_loading/test_demo_datasets.py
@@ -6,8 +6,9 @@ import pytest
 import requests
 from starlette import status
 
-from now.constants import AVAILABLE_DATASET, DEMO_DATASET_DOCARRAY_VERSION, Modalities
+from now.constants import DEMO_DATASET_DOCARRAY_VERSION, Modalities
 from now.data_loading.utils import get_dataset_url
+from now.demo_data import AVAILABLE_DATASET
 
 
 @pytest.mark.parametrize(
