+++ tests/integration/test_end_to_end.py
@@ -12,7 +12,8 @@ from now_common.options import NEW_CLUSTER
 
 from now.cli import _get_kind_path, _get_kubectl_path, cli
 from now.cloud_manager import create_local_cluster
-from now.constants import JC_SECRET, Apps, DatasetTypes, DemoDatasetNames, Modalities
+from now.constants import JC_SECRET, Apps, DatasetTypes, Modalities
+from now.demo_data import DemoDatasetNames
 from now.deployment.deployment import cmd, list_all_wolf, terminate_wolf
 from now.run_all_k8s import get_remote_flow_details
