+++ tests/unit/data_loading/test_data_loading.py
@@ -10,8 +10,9 @@ from now.apps.music_to_music.app import MusicToMusic
 from now.apps.text_to_image.app import TextToImage
 from now.apps.text_to_text.app import TextToText
 from now.apps.text_to_text_and_image.app import TextToTextAndImage
-from now.constants import DatasetTypes, DemoDatasetNames
+from now.constants import DatasetTypes
 from now.data_loading.data_loading import _load_tags_from_json_if_needed, load_data
+from now.demo_data import DemoDatasetNames
 from now.now_dataclasses import UserInput
